python/requirements.txt
python-dotenv>=1.0.0
httpx>=0.25.0
cryptography>=41.0.0 